#1066: Fixed "dev.langchain4j.exception" split package
