+++ langchain4j/src/main/java/dev/langchain4j/service/output/JsonSchemas.java
@@ -12,7 +12,7 @@ import java.lang.reflect.Type;
 import java.util.LinkedHashMap;
 import java.util.Optional;
 
-import static dev.langchain4j.exception.IllegalConfigurationException.illegalConfiguration;
+import static dev.langchain4j.service.IllegalConfigurationException.illegalConfiguration;
 import static dev.langchain4j.model.chat.request.json.JsonSchemaElementHelper.jsonObjectOrReferenceSchemaFrom;
 import static dev.langchain4j.service.TypeUtils.getRawClass;
 import static dev.langchain4j.service.TypeUtils.resolveFirstGenericParameterClass;
