+++ langchain4j/src/test/java/dev/langchain4j/service/AiServicesBuilderTest.java
@@ -2,7 +2,6 @@ package dev.langchain4j.service;
 
 import dev.langchain4j.agent.tool.Tool;
 import dev.langchain4j.data.message.AiMessage;
-import dev.langchain4j.exception.IllegalConfigurationException;
 import dev.langchain4j.model.chat.ChatLanguageModel;
 import dev.langchain4j.model.chat.mock.ChatModelMock;
 import dev.langchain4j.model.output.Response;
